multi: rename ledgeractor package to ledger
Rename the package from ledgeractor to ledger for cleaner
qualified names: ledger.LedgerActor, ledger.LedgerEntry,
ledger.UTXOAuditEntry, ledger.LedgerStore, etc.

Update all imports in db/, darepod/, and documentation files.
No functional changes.
