+++ db/utxo_audit_store_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/btcsuite/btclog/v2"
 	"github.com/lightninglabs/darepo-client/db/sqlc"
-	"github.com/lightninglabs/darepo-client/ledgeractor"
+	"github.com/lightninglabs/darepo-client/ledger"
 	"github.com/stretchr/testify/require"
 )
 
@@ -67,7 +67,7 @@ func TestUTXOAuditEnumsSeeded(t *testing.T) {
 		for _, event := range events {
 			seed++
 			err := store.InsertUTXOAuditEntry(
-				ctx, ledgeractor.UTXOAuditEntry{
+				ctx, ledger.UTXOAuditEntry{
 					OutpointHash:  makeOutpoint(seed),
 					OutpointIndex: 0,
 					AmountSat:     10_000,
@@ -103,7 +103,7 @@ func TestUTXOAuditFKRejection(t *testing.T) {
 
 	// Unknown classification should be rejected.
 	err := store.InsertUTXOAuditEntry(
-		ctx, ledgeractor.UTXOAuditEntry{
+		ctx, ledger.UTXOAuditEntry{
 			OutpointHash:  makeOutpoint(1),
 			OutpointIndex: 0,
 			AmountSat:     1000,
@@ -120,7 +120,7 @@ func TestUTXOAuditFKRejection(t *testing.T) {
 
 	// Unknown event should be rejected.
 	err = store.InsertUTXOAuditEntry(
-		ctx, ledgeractor.UTXOAuditEntry{
+		ctx, ledger.UTXOAuditEntry{
 			OutpointHash:  makeOutpoint(2),
 			OutpointIndex: 0,
 			AmountSat:     1000,
@@ -146,7 +146,7 @@ func TestUTXOAuditInsertAndList(t *testing.T) {
 	now := time.Now().Unix()
 
 	// Insert three entries at ascending timestamps.
-	entries := []ledgeractor.UTXOAuditEntry{
+	entries := []ledger.UTXOAuditEntry{
 		{
 			OutpointHash:  makeOutpoint(1),
 			OutpointIndex: 0,
@@ -209,7 +209,7 @@ func TestUTXOAuditListByBlock(t *testing.T) {
 	now := time.Now().Unix()
 
 	// Block 100 gets 2 entries, block 200 gets 1.
-	inserts := []ledgeractor.UTXOAuditEntry{
+	inserts := []ledger.UTXOAuditEntry{
 		{
 			OutpointHash:  makeOutpoint(1),
 			OutpointIndex: 0,
@@ -277,7 +277,7 @@ func TestUTXOAuditListByClassification(t *testing.T) {
 		"deposit", "deposit", "change", "deposit", "unknown",
 	} {
 		err := store.InsertUTXOAuditEntry(
-			ctx, ledgeractor.UTXOAuditEntry{
+			ctx, ledger.UTXOAuditEntry{
 				OutpointHash:  makeOutpoint(byte(i + 1)),
 				OutpointIndex: 0,
 				AmountSat:     int64((i + 1) * 1000),
@@ -337,7 +337,7 @@ func TestUTXOAuditSpendLifecycle(t *testing.T) {
 
 	// Create.
 	err := store.InsertUTXOAuditEntry(
-		ctx, ledgeractor.UTXOAuditEntry{
+		ctx, ledger.UTXOAuditEntry{
 			OutpointHash:  outpoint,
 			OutpointIndex: 3,
 			AmountSat:     1_000_000,
@@ -351,7 +351,7 @@ func TestUTXOAuditSpendLifecycle(t *testing.T) {
 
 	// Spend some blocks later.
 	err = store.InsertUTXOAuditEntry(
-		ctx, ledgeractor.UTXOAuditEntry{
+		ctx, ledger.UTXOAuditEntry{
 			OutpointHash:  outpoint,
 			OutpointIndex: 3,
 			AmountSat:     1_000_000,
