+++ ledger/AGENTS.md
@@ -1,4 +1,4 @@
-# ledgeractor
+# ledger
 
 ## Purpose
