+++ ledger/actor.go
@@ -1,4 +1,4 @@
-package ledgeractor
+package ledger
 
 import (
 	"context"
