+++ ledger/log.go
@@ -1,4 +1,4 @@
-package ledgeractor
+package ledger
 
 // Subsystem defines the logging code for the ledger actor.
 const Subsystem = "LDGR"
